+++ server_test.go
@@ -46,6 +46,8 @@ func TestServerFlags(t *testing.T) {
 	assert.True(t, names["host"])
 	assert.True(t, names["port"])
 	assert.True(t, names["database-dsn"])
+	assert.True(t, names["storage-dsn"])
+	assert.True(t, names["media-url-prefix"])
 
 	// App flags present.
 	assert.True(t, names["flaggy-key"])